Added custom PID support per image
Now the `data-pid` attribute may be specified on the `img` tag.

ISSUES:
1. Currently works whether or not the `galleryPIDs: true` option is set.
Not sure if that's correct or not.
2. If `galleryPIDs` option IS set to `true`, then the page should
automatically open the image with the custom PID in the URI, but this is
not working properly yet. Only default PIDs will open. If a custom PID
is in the URI when the page loads, as soon as the user clicks on ANY
gallery image, the image with that custom PID will open, rather than the
one the user clicked on.